requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
audioread==2.1.8
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
electrostatics==0.2.0
entrypoints==0.3
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.8
imageio==2.9.0
importlib-metadata==2.0.0
install==1.3.4
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numpy==1.19.4
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opencv-python==3.4.3.18
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pooch==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyqtgraph==0.11.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.0.5
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
Send2Trash==1.5.0
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
torch==1.7.0
torchvision==0.2.1
tornado==6.0.4
tqdm==4.51.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.24.3
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==2.0.1
wrapt==1.12.1
youtube-dl==2020.9.20
zipp==3.3.0